fix error during genotype import

+++ src/app/ui/mainwindow.cpp
@@ -75,15 +75,16 @@ void MainWindow::importFile()
                                                     QDir::homePath(),
                                                     tr("Images (*.vcf)"));
 
-    QFileInfo info(filename);
-    QString dbPath = info.dir().filePath(info.baseName()+".db");
+    QString dbPath = filename + ".db";
 
-    int ret = QMessageBox::warning(this, "Warning", "one database already exists. Do you want to remove it ?",QMessageBox::Ok|QMessageBox::Cancel );
-    if (ret == QMessageBox::Ok)
-        QFile::remove(dbPath);
-    else
-        return;
-
+    if (QFile::exists(dbPath))
+    {
+        int ret = QMessageBox::warning(this, "Warning", "one database already exists. Do you want to remove it ?",QMessageBox::Ok|QMessageBox::Cancel );
+        if (ret == QMessageBox::Ok)
+            QFile::remove(dbPath);
+        else
+            return;
+    }
 
 
     mProject->setDatabasePath(dbPath);
